requirements.txt
Django==1.11.5
django-ipware==2.1.0
fpdf==1.7.2
langdetect==1.0.7
lxml==3.8.0
mysql==0.0.1
MySQL-python==1.2.5
Markdown==2.6.9
openpyxl==2.4.9
Pillow==4.3.0
PyLTI==0.6.0
PyPDF2==1.26.0
python-docx==0.8.6
python-pptx==0.6.6

djangosaml2==0.16.1
django-admin-external-auth==0.1.0
